feat(gui): implement settings import/export system
