+++ frontend/src/components/buttons/index.ts
@@ -1,8 +1,8 @@
 // @index('./*', f => `export * from '${f.path}'`)
+export * from './backup_btn';
 export * from './convert_btn';
 export * from './import_lang_btn';
-export * from './log_dir_btn';
-export * from './log_file_btn';
+export * from './log_btn';
 export * from './path_selector';
 export * from './remove_oar_btn';
 export * from './unhide_dar_btn';
