+++ test/sample_scripts/custom_translation.js
@@ -2,6 +2,11 @@
   // Origin: https://github.com/SARDONYX-sard/dar-to-oar/blob/main/locales/en-US.json
   const i18n = {
     'all-clear-btn': 'All Clear',
+    'backup-export-btn-name': 'Export',
+    'backup-export-tooltip': 'Export current settings',
+    'backup-export-success': 'Settings exported.',
+    'backup-import-btn-name': 'Import',
+    'backup-import-tooltip': 'Import settings from Json file.',
     'conversion-complete': 'Conversion Complete.',
     'convert-btn': 'Convert',
     'convert-form-author-name': 'Mod Author Name',
@@ -49,6 +54,13 @@
     'log-level-list-tooltip3': ' Info: Log the conversion time.',
     'log-level-list-tooltip4': 'Error: Logs nothing but errors.',
     'mapping-wiki-url-leaf': 'wiki#what-is-the-mapping-file',
+    'notice-position-bottom-center': 'Bottom Center',
+    'notice-position-bottom-left': 'Bottom Left',
+    'notice-position-bottom-right': 'Bottom Right',
+    'notice-position-list-label': 'Notice Position',
+    'notice-position-top-center': 'Top Center',
+    'notice-position-top-left': 'Top Left',
+    'notice-position-top-right': 'Top Right',
     'open-log-btn': 'Open log',
     'open-log-dir-btn': 'Log(dir)',
     'open-log-dir-tooltip': 'Open the log storage location.',
@@ -67,7 +79,8 @@
       'Note:  More than twice the processing speed can be expected, but the concurrent processing results in thread termination timings being out of order, so log writes will be out of order as well, greatly reducing readability of the logs.',
     'run-parallel-label': 'Run Parallel',
     'select-btn': 'Select',
-    'tab-label-edior': 'Editor / Preset',
+    'tab-label-backup': 'Backup',
+    'tab-label-editor': 'Editor / Preset',
     'tab-label-lang': 'Language',
     'tab-label-notice': 'Notice',
     'unhide-dar-btn': 'Unhide DAR',
